Prevent toast close button from wrapping, and standardize toast card.
