+++ packages/ui/components/index.tsx
@@ -6,13 +6,13 @@ export * from './ProposalDetails'
 export * from './ProposalStatus'
 export * from './StakingModal'
 export * from './StatusIcons'
+export * from './toasts'
 
 export * from './Breadcrumbs'
 export * from './CopyToClipboard'
 export * from './CosmosMessageDisplay'
 export * from './ErrorBoundary'
 export * from './ErrorPage'
-export * from './ErrorToast'
 export * from './Execute'
 export * from './FormCard'
 export * from './GradientWrapper'
@@ -30,7 +30,6 @@ export * from './Progress'
 export * from './ProposalsInfo'
 export * from './RotatableLogo'
 export * from './SearchBar'
-export * from './SuccessToast'
 export * from './SuspenseLoader'
 export * from './TokenInfoDisplay'
 export * from './Tooltip'
