+++ packages/ui/components/toasts/index.tsx
@@ -1,0 +1,2 @@
+export * from './ErrorToast'
+export * from './SuccessToast'
